+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- cancelPayment now checks MP status, to correctly decide if we shold cancel (an authorized payment) or refund (a captured payment)
+
+### Fixed
+- Corrected getIdOrThrow utility function, which was causing problems due to incorrect type check
+
+## 0.2.4 - 2025-05-13
 ### Fixed
 - Correct response body of create payment endpoint